requirements.txt
# Python AI Agent for Natural-Language Drupal Management
# Modular Architecture Requirements

# Core CLI framework
typer>=0.9.0

# HTTP requests for JSON:API and GraphQL
requests>=2.31.0

# Environment configuration
python-dotenv>=1.0.0

# AI providers
openai>=1.0.0          # OpenAI API
anthropic>=0.7.0       # Claude API

# JSON handling and parsing
pydantic>=2.0.0

# Testing framework
pytest>=7.4.0
pytest-mock>=3.11.0

# Code formatting and linting (development)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For enhanced table output formatting
rich>=13.0.0
tabulate>=0.9.0

# Optional: For configuration file handling beyond .env
pyyaml>=6.0.0
